% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_latex.R
\name{escape_latex}
\alias{escape_latex}
\alias{escape_latex.default}
\title{Escape special characters in text intended to go into a LaTeX document}
\usage{
escape_latex(x, ...)

\method{escape_latex}{default}(x, ...)
}
\arguments{
\item{x}{The character string to escape}

\item{...}{Passed to methods}
}
\value{
`x` with LaTeX special characters escaped
}
\description{
Escape special characters in text intended to go into a LaTeX document
}
\details{
Do not escape text before sending it to this function.
}
