% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{latex_command}
\alias{latex_command}
\title{Generate a LaTeX command}
\usage{
latex_command(x, required_args = list(), optional_args = list(),
  brace_protect = TRUE)
}
\arguments{
\item{x}{The contents of the environment, a scalar vector where it will be
collapsed with newlines.}

\item{required_args}{A list of the required arguments to the environment, in
order.}

\item{optional_args}{A named list of the optional arguments to the
environment, in order.  It is an error for an optional argument to be NULL
prior to a non-optional argument.}

\item{brace_protect}{If there are no \code{required_args}, should empty
braces, \code{\\{\\}}, be added to the end of the command?}
}
\value{
A character scalar of the command
}
\description{
Generate a LaTeX command
}
\examples{
latex_command("newpar")
latex_command("newpar", brace_protect=FALSE)
latex_command("section", required_args="My first section")
latex_command("frac", required_args=list(1, 2))
}
