% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{latex_environment}
\alias{latex_environment}
\title{Generate a LaTeX environment}
\usage{
latex_environment(
  x,
  environment_name,
  required_args = list(),
  optional_args = list()
)
}
\arguments{
\item{x}{The contents of the environment, a scalar vector where it will be
collapsed with newlines.}

\item{environment_name}{The name of the environment (like "tabular" or
"minipage")}

\item{required_args}{A list of the required arguments to the environment, in
order.}

\item{optional_args}{A named list of the optional arguments to the
environment, in order.  It is an error for an optional argument to be NULL
prior to a non-optional argument.}
}
\value{
A character scalar of the environment with newlines as "\\n"
  characters embedded.
}
\description{
Generate a LaTeX environment
}
