% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minipage.R
\name{latex_minipage}
\alias{latex_minipage}
\title{Generate a LaTeX minipage environment}
\usage{
latex_minipage(
  x,
  width = NULL,
  position = NULL,
  height = NULL,
  inner_position = NULL
)
}
\arguments{
\item{x}{The text to go within the minipage.  See Details.}

\item{width}{The width of the minipage ("\\columnwidth" is a useful choice.)}

\item{position}{The position of the minipage with respect to the line where
it appears.}

\item{height}{(optional) the height of the minipage}

\item{inner_position}{The position of the contents within the minipage (only
available if \code{height} is also provided).}
}
\value{
Either a character scalar (if \code{x} is character input) or a
  character vector (if \code{x} is a list of character vectors).
}
\description{
Generate a LaTeX minipage environment
}
\details{
If \code{x} is a character vector, all text will be inserted as
  separate paragraphs (with an empty line between).  If \code{x} is a list of
  character vectors, then multiple minipages will be generated (via sapply).
  The latter is useful for building tables of minipages, for instance.
}
