% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{latex_optional_arg}
\alias{latex_optional_arg}
\title{Generate optional arguments to a LaTeX call.}
\usage{
latex_optional_arg(x = NULL, ...)
}
\arguments{
\item{x}{A character scalar (or something that can be coerced into a
character scalar)}

\item{...}{Named arguments that must be present for the optional argument to
work.}
}
\value{
The character scalar in square brackets, "[x]".
}
\description{
Generate optional arguments to a LaTeX call.
}
\examples{
latex_optional_arg(NULL)
latex_optional_arg(x="A")
}
\seealso{
Other Latex Arguments: \code{\link{latex_required_arg}}
}
\concept{Latex Arguments}
