% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{latex_required_arg}
\alias{latex_required_arg}
\title{Generate a required or optional arguments to a LaTeX call.}
\usage{
latex_required_arg(x)
}
\arguments{
\item{x}{A character scalar (or something that can be coerced into a
character scalar)}
}
\value{
The character scalar in French braces, "{x}".
}
\description{
Generate a required or optional arguments to a LaTeX call.
}
\examples{
latex_required_arg("A")
}
\seealso{
Other Latex Arguments: \code{\link{latex_optional_arg}}
}
\concept{Latex Arguments}
